test: laying the groundwork for http3 protocol test
